requirements.txt
# AeroRisk - Predictive Safety Risk Analytics Platform
# Requirements for pip installation

# ============================================
# Core Data Processing
# ============================================
pandas>=2.1.0
numpy>=1.26.0
polars>=0.20.0

# ============================================
# Database
# ============================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# ============================================
# Machine Learning
# ============================================
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.2.0
shap>=0.44.0

# ============================================
# API Framework
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# Dashboard
# ============================================
streamlit>=1.30.0
plotly>=5.18.0
altair>=5.2.0

# ============================================
# Pipeline Orchestration
# ============================================
prefect>=2.14.0

# ============================================
# HTTP & API Clients
# ============================================
httpx>=0.26.0
requests>=2.31.0

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
loguru>=0.7.2
python-multipart>=0.0.6
pyyaml>=6.0.1
python-dateutil>=2.8.2

# ============================================
# Development Dependencies
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0

# Note: Jupyter/JupyterLab removed due to Windows Long Path issues
# Install separately if needed: pip install jupyter jupyterlab
